+++ bfg9000/driver.py
@@ -1,14 +1,13 @@
 import os
+import subprocess
 import sys
 
-from . import build
-from . import log
-from . import path
+from . import build, log, path
+from .app_version import version
 from .arguments import parser as argparse
 from .backends import list_backends
 from .environment import Environment, EnvVersionError
 from .platforms.target import platform_info
-from .app_version import version
 
 logger = log.getLogger(__name__)
 
@@ -47,11 +46,11 @@ You find yourself standing at Doom's gate.
 
 
 def handle_reload_exception(e, suggest_rerun=False):
-    msg = 'Unable to reload environment'
+    msg = 'unable to reload environment'
     if str(e):
         msg += ': {}'.format(str(e))
     if suggest_rerun and isinstance(e, EnvVersionError):
-        msg += '\n  Please re-run bfg9000 manually'
+        msg += '\n  please re-run bfg9000 manually'
     logger.error(msg, exc_info=True)
     return e.code if isinstance(e, build.ScriptExitError) else 1
 
@@ -239,6 +238,26 @@ def env(parser, subparser, args, extra):
         return handle_reload_exception(e)
 
 
+def run(parser, subparser, args, extra):
+    if extra:
+        subparser.error('unrecognized arguments: {}'.format(' '.join(extra)))
+
+    if build.is_srcdir(args.builddir):
+        subparser.error('build directory must not contain a {} file'
+                        .format(build.bfgfile))
+
+    try:
+        env = Environment.load(args.builddir.string())
+        cmd = args.args
+        if cmd[0] == '--':
+            cmd = cmd[1:]
+
+        variables = env.variables.initial if args.initial else env.variables
+        return subprocess.run(cmd, env=variables).returncode
+    except Exception as e:
+        return handle_reload_exception(e)
+
+
 def e1m1(parser, subparser, args, extra):  # pragma: no cover
     import e1m1
     try:
@@ -295,11 +314,23 @@ def main():
     )
     env_p.set_defaults(func=env, parser=env_p)
     env_p.add_argument('-u', '--unique', action='store_true',
-                       help='only show variables that differ from the ' +
-                       'current environment')
+                       help=('only show variables that differ from the ' +
+                             'current environment'))
     env_p.add_argument('builddir', type=argparse.Directory(must_exist=True),
                        metavar='BUILDDIR', nargs='?', default='.',
                        help='build directory')
+
+    run_p = subparsers.add_parser(
+        'run', description=env_desc, help='FIXME'
+    )
+    run_p.set_defaults(func=run, parser=run_p)
+    run_p.add_argument('-I', '--initial', action='store_true',
+                       help='use initial environment variable state')
+    run_p.add_argument('-B', '--builddir',
+                       type=argparse.Directory(must_exist=True),
+                       metavar='BUILDDIR', default='.', help='build directory')
+    run_p.add_argument('args', metavar='ARGS', nargs=argparse.REMAINDER,
+                       help='command argument')
 
     e1m1_p = subparsers.add_parser('e1m1', description=e1m1_desc)
     e1m1_p.set_defaults(func=e1m1, parser=e1m1_p)