+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## v0.7.0 (in progress)
 
 ### New features
+- Add `bfg9000 run` command to allow executing other commands using the
+  environment variables for a given build
 - `pkg-config` executable can now be found by guessing based on other compilers
 
 ### Breaking changes